Bill composite created h files
I also cleaned up my use of the singleton and completed the logic for tabs and bill payment simulation.
What is left for me to do now is to complete the cpp files for the bill composite.
I will then possibly be safe to log a merge request with dev branch.